# Conflicts:
#	AuthenticateService/src/main/resources/logback.log
#	AuthenticateService/src/main/resources/mappers/UserMapper.xml
#	AuthenticateService/target/classes/cn/edu/sicnu/cs/dao/WorkordersMapper.class
#	AuthenticateService/target/generated-sources/annotations/cn/edu/sicnu/cs/datamapper/LogSmallMapperImpl.java

+++ AuthenticateService/src/main/java/cn/edu/sicnu/cs/dao/TaskMapper.java
@@ -36,7 +36,7 @@ public interface TaskMapper {
 
     int updateByPrimaryKey(Task record);
 
-    int findSelfTaskCntByStatus(long uid, String status);
+    long findSelfTaskCntByStatus(long uid, String status);
